refactor(HeadContent): made class, with static methods
HeadContent class is a class with the addComponentStyleSheet and addHeadContent functions as static methods. It replaces those files. The class is used for anything to do with creating/modifying the head content.

BREAKING CHANGE: The codebase still uses the old methods, not the class yet. Because of that, it will give errors.