refactor(GlobalStyles): changed setglobalstyles function to be a class instead
Made a class with two private methods, that are used in the class only. There is a set globalStyles property to set the globalStyles if the user wish to do that later on. There is also a appendStyles method, where you can pass in an array of styles, which will be concatted to the already existing styles, updating the global styles.